requirements.txt

Pillow==10.2.0
protobuf==3.20.3
python-dotenv==1.0.0
streamlit==1.20.0
